Command struct, parseInput working, quit command working
both of these methods are now working with the new struct i made to
store the current command + arguments.

in this stage, each command typed at the prompt prints out a debug
message that separates the command from the arguments given.

it will parse correctly with multiple spaces, tabs, etc.