Refactoring, commenting, restructuring
everything is very readable now. all debug comments removed. edge case
tested. see the bash script “lss”—try calling /bin/echo 30 & in the
background and then calling lss in the foreground. works great.

one weird thing is that if a user decides to execute something in the
background that barely takes any time, the prompt is printed out before
the output of the background process, so things get messed up. maybe a
background process should always print a “>” right before it terminates?

still need to handle strange unexpected input from the user, and make
the status messages we print out are the same as tcsh